requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
jinja2==3.1.2
    # via -r requirements.in
markupsafe==2.1.1
    # via jinja2
pygraphviz==1.10
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
